refactor: refactor classs in repo and usecase
